+++ src/simple_sqlite3_orm/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from simple_sqlite3_orm._orm import ORMBase
+from simple_sqlite3_orm._orm import AsyncORMThreadPoolBase, ORMBase, ThreadPoolExecutor
 from simple_sqlite3_orm._sqlite_spec import (
     SQLiteBuiltInFuncs,
     SQLiteStorageClass,
@@ -25,7 +25,9 @@ __all__ = [
     "SQLiteTypeAffinityLiteral",
     "TypeAffinityRepr",
     "TableSpec",
+    "AsyncORMThreadPoolBase",
     "ORMBase",
+    "ThreadPoolExecutor",
     "DatetimeISO8601",
     "DatetimeUnixTimestamp",
     "DatetimeUnixTimestampInt",
